requirements.txt
Flask
Gunicorn
CORS
flask_cors
uvicorn
google-cloud-texttospeech
python-dotenv